Add fallback language based on language
Fallback language code is generated by the first two chars of installed language (>de<_CH) and this is used as suffix in upper case (de_>DE<).
Es an example, when language de_CH is installed the $fallbacklang will be set to de_DE (first two chars are used as 4. and 5. char).
When an extension does not have a de_CH.xxx.ini file, the de_DE.xxx.ini file will be used.