Fallback Language import
Fallback language code is generated by the first two chars of installed language (>de<-CH) and this is used as suffix in upper case (de->DE<).
Es an example, when language de-CH is installed the $fallbackLang will be set to de-DE (first two chars are used as 4. and 5. char).
When an extension does not have a de-CH.xxx.ini file, the de-DE.xxx.ini file will be used.